Message , User, Notification models

+++ src/main/java/com/linkedin/database/repo/UserRepository.java
@@ -6,7 +6,13 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User,Long> {
+
+//    @Query("SELECT * FROM 'users'")
+//    List<User> findAll();
+
 
 }
